% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_builder.R
\name{build_gene_disease_reference}
\alias{build_gene_disease_reference}
\title{Aggregate all the databases needed to run PhenolyzerR}
\usage{
build_gene_disease_reference()
}
\value{
A Tibble
}
\description{
`build_gene_disease_reference` is downloads gene-disease mappings from OMIM, ClinVar, and the GWAS repository.
This will only need to be run once. When the script finishes running,
you will have a table called `DB_COMPILED_GENE_DISEASE_SCORES` in your local environment.
}
