% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_builder.R
\name{build_gene_id_syn}
\alias{build_gene_id_syn}
\title{This function downloads a list of all genes and saves a tibble of Entrez gene IDs, gene number,
and synonyms in the global environment}
\usage{
build_gene_id_syn()
}
\description{
This function downloads a list of all genes and saves a tibble of Entrez gene IDs, gene number,
and synonyms in the global environment
}
