% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_gene_prioritization.R
\name{disease_extension}
\alias{disease_extension}
\title{This function uses the public ontology_search function to try both databases}
\usage{
disease_extension(term, exact_match = FALSE)
}
\description{
This function uses the public ontology_search function to try both databases
}
