% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_predicted_genes.R
\name{output_predicted_genes}
\alias{output_predicted_genes}
\title{Predict disease genes from a seed gene list}
\usage{
output_predicted_genes(seed_gene_list, n_seeds = 10)
}
\arguments{
\item{seed_gene_list}{The tibble outputted by 'output_gene_prioritization'}

\item{n_seeds}{The number of seed genes to use to extend genes. Defaults to 10}
}
\value{
A tibble
}
\description{
This function utlizes the structure of the Gene Ontology resource to predict
unreported disease genes from a list of seed genes. It accomplishes this task
by finding the most granular GO term that a particular gene belongs to, and
collecting all other genes belonging to that term. It scores predicted genes
based on their frequencies in the related GO terms.
}
\details{
By default, only 10 seed genes are used to predict new genes. This can be changed, but increasing
this parameter will increase runtime exponentially
}
\examples{
autism_seed_genes <- output_gene_prioritization("autism")
output_predicted_genes(autism_seed_genes)

}
