% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_predicted_genes.R
\name{retrieve_go_information}
\alias{retrieve_go_information}
\alias{retrieve_go_OBO}
\title{Retrieve GO Information}
\usage{
retrieve_go_information()

retrieve_go_OBO()
}
\value{
NULL, Puts variables GO_GAF and GO in your local environment
}
\description{
These function download the GO OBO file and GO annotation file, which are used by the \code{output_predicted_genes}
function.
}
