% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_search.R
\name{retrieve_ontologies}
\alias{retrieve_ontologies}
\title{Retrieve the obo ontology files needed for disease term descendant searching.
Currently, this package loads the CTD Disease ontology and the Disease ongology (doid.obo)}
\usage{
retrieve_ontologies()
}
\description{
Retrieve the obo ontology files needed for disease term descendant searching.
Currently, this package loads the CTD Disease ontology and the Disease ongology (doid.obo)
}
