% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_gene_prioritization.R
\name{write_gene_list}
\alias{write_gene_list}
\title{Write an itemized gene list to file}
\usage{
write_gene_list(itemized_gene_list, filename)
}
\arguments{
\item{itemized_gene_list}{The output of list \code{output_gene_prioritization} with \code{output_mode == "list"}}

\item{filename}{File name to write to the the filesystem}
}
\description{
If you want to see the sources for each gene call, this function will write the results
of \code{output_gene_prioritization} with \code{output_mode == "list"} to a text file that is conveniently
readable.
}
