% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_builder.R
\name{build_gene_id_syn}
\alias{build_gene_id_syn}
\title{This function pulls database information from the web and integrates it to one table for use in the score_genes function
This should do everything the integrate_gene_disease.pl
and the compile_gene_disease_score.pl script does in Phenolyzer}
\usage{
build_gene_id_syn()
}
\description{
This function pulls database information from the web and integrates it to one table for use in the score_genes function
This should do everything the integrate_gene_disease.pl
and the compile_gene_disease_score.pl script does in Phenolyzer
}
