% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_gene_prioritization.R
\name{output_gene_prioritization}
\alias{output_gene_prioritization}
\title{Retrieve genes related to a disease query}
\usage{
output_gene_prioritization(disease_term, ontologies = c("ctd", "doid"))
}
\arguments{
\item{disease_term}{A disease for which to find related genes}

\item{ontologies}{Defaults to c("ctd", "doid").}
}
\description{
This function calls \code{disease_extension} to perform a descendant search, then prioritizes the
disease terms matched to the disease-gene databases using system described in the paper.
}
\details{
In brief, each database has its own quality metrics. These are given an ad hoc score, and matches
are weighted based on this score. The prioritized list is ordered, and normalized by diving through
all scores by the max score.
}
\examples{
output_gene_prioritization("alzheimer's disease")
output_gene_prioritization("alzheimer's disease", ontologies = "doid")
}
