% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_search.R
\name{retrieve_ontologies}
\alias{retrieve_ontologies}
\title{Retrieve ontology files from the internet}
\usage{
retrieve_ontologies()
}
\description{
retrieve_ontologies() adds \code{doid} and \code{ctd} to the local environment
}
